Dvaraka (#28)
* upgrade gradle from 3.3 to 8.1
* upgrade to lib-aeapps v5 alpha
* upgrade agp
* target android api 33
* update travis config

+++ app/src/main/java/com/ae/apps/tripmeter/views/adapters/TripRecyclerViewAdapter.java
@@ -19,8 +19,11 @@
  */
 package com.ae.apps.tripmeter.views.adapters;
 
-import android.support.v7.widget.PopupMenu;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.appcompat.widget.PopupMenu;
+import androidx.recyclerview.widget.RecyclerView;
+
+import android.annotation.SuppressLint;
 import android.view.LayoutInflater;
 import android.view.MenuInflater;
 import android.view.MenuItem;
@@ -56,6 +59,7 @@ public class TripRecyclerViewAdapter extends RecyclerView.Adapter<TripRecyclerVi
         mListUpdateListener = listUpdateListener;
     }
 
+    @NonNull
     @Override
     public ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
         View view = LayoutInflater.from(parent.getContext())
@@ -65,7 +69,7 @@ public class TripRecyclerViewAdapter extends RecyclerView.Adapter<TripRecyclerVi
     }
 
     @Override
-    public void onBindViewHolder(final ViewHolder holder, final int position) {
+    public void onBindViewHolder(final ViewHolder holder, @SuppressLint("RecyclerView") final int position) {
         Trip trip = mValues.get(position);
         holder.mItem = trip;
         holder.mTripName.setText(trip.getName());
@@ -113,7 +117,7 @@ public class TripRecyclerViewAdapter extends RecyclerView.Adapter<TripRecyclerVi
         return mValues.size();
     }
 
-    class ViewHolder extends RecyclerView.ViewHolder {
+    static class ViewHolder extends RecyclerView.ViewHolder {
         final View mView;
         final TextView mTripName;
         final TextView mTripDate;
@@ -125,12 +129,13 @@ public class TripRecyclerViewAdapter extends RecyclerView.Adapter<TripRecyclerVi
             super(view);
             view.setClickable(true);
             mView = view;
-            mTripName = (TextView) view.findViewById(R.id.tripName);
-            mTripDate = (TextView) view.findViewById(R.id.tripDate);
-            mTripMemberCount = (TextView) view.findViewById(R.id.tripMemberCount);
-            mPopupMenu = (ImageView) view.findViewById(R.id.tripMenu);
+            mTripName =  view.findViewById(R.id.tripName);
+            mTripDate =  view.findViewById(R.id.tripDate);
+            mTripMemberCount =  view.findViewById(R.id.tripMemberCount);
+            mPopupMenu =  view.findViewById(R.id.tripMenu);
         }
 
+        @NonNull
         @Override
         public String toString() {
             return super.toString() + " '" + mTripDate.getText() + "'";
@@ -139,7 +144,7 @@ public class TripRecyclerViewAdapter extends RecyclerView.Adapter<TripRecyclerVi
 
     private class TripMenuItemClickListener implements PopupMenu.OnMenuItemClickListener {
 
-        private int mPosition;
+        private final int mPosition;
 
         TripMenuItemClickListener(int position) {
             this.mPosition = position;
@@ -147,13 +152,12 @@ public class TripRecyclerViewAdapter extends RecyclerView.Adapter<TripRecyclerVi
 
         @Override
         public boolean onMenuItemClick(MenuItem item) {
-            switch (item.getItemId()) {
-                case R.id.action_delete_trip:
-                    mListUpdateListener.deleteTrip(mValues.get(mPosition));
-                    return true;
-                case R.id.action_edit_trip_name:
-                    mListUpdateListener.updateTrip(mValues.get(mPosition));
-                    return true;
+            if(item.getItemId() == R.id.action_delete_trip) {
+                mListUpdateListener.deleteTrip(mValues.get(mPosition));
+                return true;
+            } else if (item.getItemId() == R.id.action_edit_trip_name) {
+                mListUpdateListener.updateTrip(mValues.get(mPosition));
+                return true;
             }
             return false;
         }
